feat(tests): extract createArrayOfNumbers to module
